backend/requirements.txt
anyio==4.12.0
bcrypt==5.0.0
certifi==2025.11.12
cffi==2.0.0
click==8.3.1
colorama==0.4.6
cryptography==46.0.3
dnspython==2.8.0
ecdsa==0.19.1
fastapi==0.95.0
h11==0.14.0
httpcore==0.17.3
httptools==0.7.1
httpx==0.24.1
idna==3.11
iniconfig==2.3.0
motor==3.7.1
packaging==25.0
passlib==1.7.4
pluggy==1.6.0
pyasn1==0.6.1
pycparser==2.23
pydantic==1.10.12
pymongo==4.15.5
pytest==7.4.0
python-dotenv==1.0.0
python-jose==3.5.0
PyYAML==6.0.3
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
starlette==0.26.1
typing_extensions==4.15.0
uvicorn==0.22.0
watchfiles==1.1.1
websockets==15.0.1
